+++ app/src/main/java/algorithm/master/currencyconverterpro/domain/model/converter/ConverterCurrencyModel.kt
@@ -4,10 +4,10 @@ package algorithm.master.currencyconverterpro.domain.model.converter
  * Created by promasterguru on 15/11/2022.
  */
 data class ConverterCurrencyModel(
-    val amount: Double,
+    val amount: Float,
     val from: String,
     val to: String,
-    val rate: Double,
+    val rate: Float,
     val date: String,
-    val result: Double
+    val result: Float
 )
